Extracted nested header wrapper classes to full class status for use as a data transfer object.
