+++ src/Kabomu/Mediator/Handling/DefaultContextRequest.cs
@@ -15,7 +15,7 @@ namespace Kabomu.Mediator.Handling
         {
             RawRequest = rawRequest ?? throw new ArgumentNullException(nameof(rawRequest));
             Environment = requestEnvironment ?? new Dictionary<string, object>();
-            Headers = new DefaultHeadersWrapper(rawRequest);
+            Headers = new QuasiHttpHeadersWrapper(() => rawRequest.Headers, null);
             _registry = new DefaultMutableRegistry();
         }
 
@@ -62,31 +62,6 @@ namespace Kabomu.Mediator.Handling
         public (bool, object) TryGetFirst(object key, Func<object, (bool, object)> transformFunction)
         {
             return _registry.TryGetFirst(key, transformFunction);
-        }
-
-        class DefaultHeadersWrapper : IHeaders
-        {
-            private readonly IQuasiHttpRequest _parent;
-
-            public DefaultHeadersWrapper(IQuasiHttpRequest parent)
-            {
-                _parent = parent;
-            }
-
-            public string Get(string name)
-            {
-                var rawHeaders = _parent.Headers;
-                IList<string> values = null;
-                if (rawHeaders != null && rawHeaders.ContainsKey(name))
-                {
-                    values = rawHeaders[name];
-                }
-                if (values != null && values.Count > 0)
-                {
-                    return values[0];
-                }
-                return null;
-            }
         }
     }
 }
